Move transform method to a utility class
